Changing column data type

+++ db/migrate/20200508193804_create_accounts.rb
@@ -2,7 +2,7 @@ class CreateAccounts < ActiveRecord::Migration[5.1]
   def change
     create_table :accounts do |t|
       t.integer :account
-      t.float :balance
+      t.integer :balance
 
       t.timestamps
     end
